--- lib/persistor/platform_persistor/platform_persistor.dart
@@ -1,1 +1,0 @@
-export 'native_persistor.dart' if (dart.library.html) 'web_persistor.dart';
